fix(review): move _identify_optional_output_fields to _rule_helpers.py
Eliminates cross-import between rule modules (rules_optional_capture →
campaign/rules_campaign_capture) by moving the shared helper to the
canonical _rule_helpers.py module. Re-exports from the original location
for backward compatibility.
